Add TestableRelease config to use for unit testing in Travis.
Only 'Enable Testability' setting is different from Release config. The setting prohibits optimizations that depend on not exporting internal symbols from the app or framework. To keep the optimizations for Release config and to test with a config close to the Release config as much as possible, TestableRelease config is added.