Fix Issue #10 that dependency injection to view controllers in a storyboard does not work.
The cause of the issue was that `entry.initCompleted as? (Resolvable, C) -> ()` in `Container` failed with Xcode 7.1 (beta). The cast with type `C` worked dynamically with Xcode 7.0, but not with Xcode 7.1. To avoid the problem, the type of initCompleted closure is always fixed to `(Resolvable, Controller) -> ()` by wrapping `(Resolvable, C) -> ()` type closure.

The fix is the same as the commit 63a5e0000d9d501078f75672f9972f8c0505978f, which was applied for Xcode 6.4 with Swift 1.2.